refactor: rename SimplifiedTestSupport to TestSupport
Final removal of "Simplified" naming from the codebase.
Test support utilities now have a clean, simple name.
